requirements.txt
fastapi
uvicorn
google-generativeai
Pillow
python-dotenv
python-multipart